refactor to use a common sha1 function
